backend/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
Flask-JWT-Extended
PyMySQL
python-dotenv
pandas
matplotlib